Implemented Poll Results
Seo, All the required information is available on the result.jsp page.
It has the information on Poll as well as a set of results for solved
polls.
Use this data to create graphs and information tables.
Use datatables library to display table data and D3 to create graph.
Please finish it ASAP